+++ src/organism/DirectmsgRoom.tsx
@@ -1,95 +1,18 @@
-import { useEffect, useState } from 'react';
-import { useSetRecoilState } from 'recoil';
-import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { useAutoScroll, useSocketRef } from '@/hooks/chat';
+import { useState } from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { useAutoScroll } from '@/hooks/chat';
 import { fetcher } from '@/hooks/fetcher';
-import StatusIndicator from '@/molecule/StatusIndicator';
-import { dmHistoryState, type DmInfo } from '@/states/dmHistoryState';
 import { type Friend } from '@/hooks/friends';
+import {
+  useDirectMessages,
+  useDirectMessageSocket,
+  useUpdateDmHistory,
+} from '@/hooks/dm';
+import StatusIndicator from '@/molecule/StatusIndicator';
 
 interface DirectmsgRoomProps {
   opponent: string;
   close(): void;
-}
-
-interface Message {
-  sender: Friend;
-  payload: string;
-  created: Date;
-}
-
-interface SubscribeData {
-  sender: Friend;
-  payload: string;
-}
-
-function useDirectMessages(opponent: string) {
-  const data = useQuery<Message[]>({
-    queryKey: ['dm/messages', opponent],
-    queryFn: async () => {
-      const res = await fetcher(`/dm/${opponent}/messages`);
-
-      if (res.ok) {
-        const data = await res.json();
-        return data.messages;
-      }
-      return [];
-    },
-    refetchOnWindowFocus: false,
-  });
-
-  return data;
-}
-
-function useDirectMessage(opponent: string) {
-  const queryClient = useQueryClient();
-  const socketRef = useSocketRef(`ws://${import.meta.env.VITE_BASE_URL}:9992`);
-  const { data: messages } = useDirectMessages(opponent);
-  const setDmHistory = useSetRecoilState(dmHistoryState);
-
-  useEffect(() => {
-    const socket = socketRef.current;
-
-    const handleSub = (data: SubscribeData) => {
-      queryClient.setQueryData<Message[]>(
-        ['dm/messages', opponent],
-        (prevMsg) => {
-          const newMessage: Message = {
-            sender: data.sender,
-            payload: data.payload,
-            created: new Date(),
-          };
-          return prevMsg ? [...prevMsg, newMessage] : [newMessage];
-        },
-      );
-      setDmHistory((currHistory) =>
-        currHistory.map((dmInfo) => {
-          if (dmInfo.opponent.nickname === opponent) {
-            return { ...dmInfo, last_dm: new Date(), last_read: new Date() };
-          }
-          return dmInfo;
-        }),
-      );
-    };
-
-    const handleAnnounce = (data: unknown) => {
-      console.log(data);
-    };
-
-    socket.emit('join', { opponent });
-
-    socket.on('subscribe', handleSub);
-    socket.on('subscribe_self', handleSub);
-    socket.on('announcement', handleAnnounce);
-
-    return () => {
-      socket.off('subscribe', handleSub);
-      socket.off('subscribe_self', handleSub);
-      socket.off('announcement', handleAnnounce);
-    };
-  }, [opponent, queryClient, socketRef, setDmHistory]);
-
-  return { messages, socket: socketRef.current };
 }
 
 function useOpponent(nickname: string) {
@@ -105,107 +28,118 @@ function useOpponent(nickname: string) {
 }
 
 function DirectmsgRoom({ opponent, close }: DirectmsgRoomProps) {
-  const [content, setContent] = useState('');
-  const { messages, socket } = useDirectMessage(opponent);
+  const { socket } = useDirectMessageSocket(opponent);
   const { data: opponentInfo } = useOpponent(opponent);
-  const autoScrollRef = useAutoScroll(messages);
+  useUpdateDmHistory(opponentInfo);
 
-  const setDmHistory = useSetRecoilState(dmHistoryState);
-  useEffect(() => {
-    if (opponentInfo) {
-      setDmHistory((currHistory) => {
-        const newHistory: DmInfo = {
-          opponent: opponentInfo,
-          last_read: new Date(),
-          last_dm: new Date(),
-        };
-        const idx = currHistory.findIndex(
-          (dmInfo) => dmInfo.opponent.user_id === opponentInfo.user_id,
-        );
-        return idx !== -1
-          ? [
-              ...currHistory.slice(0, idx),
-              newHistory,
-              ...currHistory.slice(idx + 1),
-            ]
-          : [...currHistory, newHistory];
-      });
+  const handleSubmit = (content: string) => {
+    content = content.trim();
+    if (content) {
+      socket.emit('publish', { opponent, payload: content });
     }
-  }, [opponentInfo, setDmHistory]);
-
-  const sendMessage = () => {
-    const msg = content.trim();
-    if (msg) {
-      socket.emit('publish', { opponent, payload: msg });
-    }
-    setContent('');
   };
 
   return (
     <>
       <div className="flex h-fit w-full items-center justify-between border-b border-inherit bg-neutral-800 p-2">
-        {opponentInfo ? (
-          <div className="flex flex-row gap-2 p-2">
-            <div className="m-1 flex h-10 w-10 items-center justify-center overflow-hidden rounded-full">
-              <img
-                src={opponentInfo.prof_img}
-                alt="profile"
-                className="h-full w-full object-cover"
-              />
-            </div>
-            <div className="flex flex-col">
-              <p>{opponentInfo.nickname}</p>
-              <div className="flex flex-row items-center space-x-2">
-                <StatusIndicator status={opponentInfo.status} />
-                <p className="min-w-0 truncate text-xs">
-                  {opponentInfo.status ?? 'unknown'}
-                </p>
-              </div>
-            </div>
-            <button className="px-3">:</button>
-          </div>
-        ) : (
-          opponent
-        )}
+        {opponentInfo ? <OpponentProfile opponent={opponentInfo} /> : opponent}
         <button onClick={close} className="px-1">
           ⬅
         </button>
       </div>
-      <div
-        ref={autoScrollRef}
-        className="h-full w-full grow overflow-y-auto border-b border-inherit"
-      >
-        <ul className="flex h-fit w-full flex-col items-start justify-start">
-          {messages?.map((message, idx) => (
-            <li
-              key={idx}
-              className="h-fit w-full break-words p-1 px-5 text-xs"
-            >{`${message.sender.nickname}: ${message.payload}`}</li>
-          ))}
-        </ul>
-      </div>
-      <div className="h-30 flex">
-        <textarea
-          placeholder="plase input here."
-          className="h-20 w-full resize-none border-none bg-neutral-300 text-black"
-          value={content}
-          onChange={(e) => setContent(e.target.value)}
-          onKeyUp={(e) => {
-            if (e.key === 'Enter') {
-              sendMessage();
-            }
-          }}
-        />
-        <button
-          className="h-full border-b border-inherit bg-neutral-800 px-3"
-          onClick={() => {
-            sendMessage();
-          }}
-        >
-          send
-        </button>
-      </div>
+      <MessageContainer opponent={opponent} />
+      <TextInput onSubmit={handleSubmit} />
     </>
+  );
+}
+
+interface OpponentProfileProps {
+  opponent: Friend;
+}
+
+function OpponentProfile({ opponent }: OpponentProfileProps) {
+  return (
+    <div className="flex flex-row gap-2 p-2">
+      <div className="m-1 flex h-10 w-10 items-center justify-center overflow-hidden rounded-full">
+        <img
+          src={opponent.prof_img}
+          alt="profile"
+          className="h-full w-full object-cover"
+        />
+      </div>
+      <div className="flex flex-col">
+        <p>{opponent.nickname}</p>
+        <div className="flex flex-row items-center space-x-2">
+          <StatusIndicator status={opponent.status} />
+          <p className="min-w-0 truncate text-xs">
+            {opponent.status ?? 'unknown'}
+          </p>
+        </div>
+      </div>
+      <button className="px-3">:</button>
+    </div>
+  );
+}
+
+interface MessageContainerProps {
+  opponent: string;
+}
+
+function MessageContainer({ opponent }: MessageContainerProps) {
+  const { data: messages } = useDirectMessages(opponent);
+  const autoScrollRef = useAutoScroll(messages);
+
+  return (
+    <div
+      ref={autoScrollRef}
+      className="h-full w-full grow overflow-y-auto border-b border-inherit"
+    >
+      <ul className="flex h-fit w-full flex-col items-start justify-start">
+        {messages?.map((message, idx) => (
+          <li
+            key={idx}
+            className="h-fit w-full break-words p-1 px-5 text-xs"
+          >{`${message.sender.nickname}: ${message.payload}`}</li>
+        ))}
+      </ul>
+    </div>
+  );
+}
+
+interface TextInputProps {
+  onSubmit(content: string): void;
+}
+
+function TextInput({ onSubmit }: TextInputProps) {
+  const [content, setContent] = useState('');
+
+  const sendMessage = () => {
+    onSubmit(content);
+    setContent('');
+  };
+
+  return (
+    <div className="h-30 flex">
+      <textarea
+        placeholder="plase input here."
+        className="h-20 w-full resize-none border-none bg-neutral-300 text-black"
+        value={content}
+        onChange={(e) => setContent(e.target.value)}
+        onKeyUp={(e) => {
+          if (e.key === 'Enter') {
+            sendMessage();
+          }
+        }}
+      />
+      <button
+        className="h-full border-b border-inherit bg-neutral-800 px-3"
+        onClick={() => {
+          sendMessage();
+        }}
+      >
+        send
+      </button>
+    </div>
   );
 }
